add lpc step
