+++ app/src/assets/localization/en/protocol_setup.json
@@ -6,8 +6,10 @@
   "extra_attention_warning_title": "Secure labware and modules before proceeding to run",
   "labware_setup_step_description": "Position full tip racks and labware in the deck slots as shown in the deck map.",
   "labware_setup_step_title": "Labware Setup",
+  "labware_position_check_step_description": "Recommended workflow that helps you verify the position of each labware on the deck.",
+  "labware_position_check_step_title": "Labware Position Check",
   "how_offset_data_works": "How labware offsets work",
-  "what_labware_offset_is": "A Labware Offset is a type of positional adjustment that accounts for small, real-world variances in the overall position of the labware on an OT-2’s deck. Labware Offset data is unique to a specific combination of labware definition, deck slot, and OT-2.",
+  "what_labware_offset_is": "A Labware Offset is a type of positional adjustment that accounts for small, real-world variances in the overall position of the labware on a robot’s deck. Labware Offset data is unique to a specific combination of labware definition, deck slot, and robot.",
   "learn_more_about_robot_cal_offset": "Learn more about Robot Calibration",
   "learn_more_about_offset_data": "Learn more about Labware Offset Data",
   "learn_how_it_works": "Learn how it works",
@@ -15,7 +17,7 @@
   "multiple_modules_help_link_title": "See How To Set Up Modules of the Same Type",
   "multiple_modules_modal": "Setting up modules of the same type",
   "labware_help_example": "<block>Alice is using a labware in Slot 6. During labware position check, she adjust the pipette position while checking the A1 of the labware to 0.2mm in X, and 1.2mm in Z. Later, Bob is preparing to run the same protocol on that robot. The labware offset that Alice created in Slot 6 will be applied for Bob’s protocol unless he changes or clears it.</block>",
-  "multiple_modules_explanation": "To use more than one of the same module in a protocol, you first need to plug in the module that’s called first in your protocol to the lowest numbered USB port on the OT-2. Continue in the same manner with additional modules.",
+  "multiple_modules_explanation": "To use more than one of the same module in a protocol, you first need to plug in the module that’s called first in your protocol to the lowest numbered USB port on the robot. Continue in the same manner with additional modules.",
   "multiple_modules_learn_more": "Learn more about using multiple modules of the same type",
   "multiple_modules_example": "Your protocol has two Temperature Modules. The Temperature Module attached to the first port starting from the left will be related to the first Temperature Module in your protocol while the second Temperature Module loaded would be related to the Temperature Module connected to the next port to the right. If using a hub, follow the same logic with the port ordering.",
   "offset_data": "Offset Data",
@@ -29,8 +31,8 @@
   "calibrate_now_cta": "Calibrate Now",
   "deck_calibration_title": "Deck Calibration",
   "last_calibrated": "Last calibrated: {{date}}",
-  "module_setup_step_description": "Plug in and turn on the required module via the OT-2 USB Port. Place the module as shown in the deck map.",
-  "module_setup_step_description_plural": "Plug in and turn on the required modules via the OT-2 USB Ports. Place the modules as shown in the deck map.",
+  "module_setup_step_description": "Plug in and turn on the required module via the robot's USB Port. Place the module as shown in the deck map.",
+  "module_setup_step_description_plural": "Plug in and turn on the required modules via the robot's USB Ports. Place the modules as shown in the deck map.",
   "module_setup_step_title": "Module Setup",
   "modules_setup_step_title": "Module Setup",
   "mount_title": "{{mount}} MOUNT:",
@@ -63,7 +65,7 @@
   "hub_port_connected": "USB Port {{port}} via hub",
   "usb_connected_no_port_info": "USB Port Connected",
   "robot_cal_help_title": "How Robot Calibration Works",
-  "robot_cal_description": "Robot calibration establishes how the OT-2 knows where it is in relation to the deck. Accurate Robot calibration is essential to run protocols successfully. Robot calibration has 3 parts: Deck calibration, Tip Length calibration and Pipette Offset calibration.",
+  "robot_cal_description": "Robot calibration establishes how the robot knows where it is in relation to the deck. Accurate Robot calibration is essential to run protocols successfully. Robot calibration has 3 parts: Deck calibration, Tip Length calibration and Pipette Offset calibration.",
   "learn_more_about_robot_cal_link": "Learn more about robot calibration",
   "deck_cal_description": "This measures the deck X and Y values relative to the gantry. Deck Calibration is the foundation for Tip Length Calibration and Pipette Offset Calibration.",
   "deck_cal_description_bullet_1": "Perform Deck Calibration during new robot setup.",
@@ -157,7 +159,6 @@
   "view_moam": "View setup instructions for placing modules of the same type to the robot.",
   "secure": "Secure",
   "additional_off_deck_labware": "Additional Off-Deck Labware",
-  "recommended_workflow_for_labware_positioning": "Recommended workflow that helps you verify the position of each labware on the deck. <anchor>Learn how it works</anchor>",
   "view_current_offsets": "View current offsets",
   "view_module_setup_instructions": "View module setup instructions",
   "deck_map": "Deck Map",
