+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunSetup.tsx
@@ -23,18 +23,21 @@ import {
 } from '../hooks'
 import { useMostRecentCompletedAnalysis } from '../../LabwarePositionCheck/useMostRecentCompletedAnalysis'
 import { SetupLabware } from './SetupLabware'
+import { SetupLabwarePositionCheck } from './SetupLabwarePositionCheck'
 import { SetupRobotCalibration } from './SetupRobotCalibration'
 import { SetupModules } from './SetupModules'
 import { SetupStep } from './SetupStep'
 import { SetupLiquids } from './SetupLiquids'
 const ROBOT_CALIBRATION_STEP_KEY = 'robot_calibration_step' as const
 const MODULE_SETUP_KEY = 'module_setup_step' as const
+const LPC_KEY = 'labware_position_check_step' as const
 const LABWARE_SETUP_KEY = 'labware_setup_step' as const
 const LIQUID_SETUP_KEY = 'liquid_setup_step' as const
 
 export type StepKey =
   | typeof ROBOT_CALIBRATION_STEP_KEY
   | typeof MODULE_SETUP_KEY
+  | typeof LPC_KEY
   | typeof LABWARE_SETUP_KEY
   | typeof LIQUID_SETUP_KEY
 
@@ -64,6 +67,7 @@ export function ProtocolRunSetup({
   )
   const [stepsKeysInOrder, setStepKeysInOrder] = React.useState<StepKey[]>([
     ROBOT_CALIBRATION_STEP_KEY,
+    LPC_KEY,
     LABWARE_SETUP_KEY,
   ])
 
@@ -77,6 +81,7 @@ export function ProtocolRunSetup({
       nextStepKeysInOrder = [
         ROBOT_CALIBRATION_STEP_KEY,
         MODULE_SETUP_KEY,
+        LPC_KEY,
         LABWARE_SETUP_KEY,
         LIQUID_SETUP_KEY,
       ]
@@ -84,11 +89,13 @@ export function ProtocolRunSetup({
       nextStepKeysInOrder = [
         ROBOT_CALIBRATION_STEP_KEY,
         MODULE_SETUP_KEY,
+        LPC_KEY,
         LABWARE_SETUP_KEY,
       ]
     } else if (showLiquidSetup) {
       nextStepKeysInOrder = [
         ROBOT_CALIBRATION_STEP_KEY,
+        LPC_KEY,
         LABWARE_SETUP_KEY,
         LIQUID_SETUP_KEY,
       ]
@@ -126,7 +133,7 @@ export function ProtocolRunSetup({
     [MODULE_SETUP_KEY]: {
       stepInternals: (
         <SetupModules
-          expandLabwareSetupStep={() => setExpandedStepKey(LABWARE_SETUP_KEY)}
+          expandLabwarePositionCheckStep={() => setExpandedStepKey(LPC_KEY)}
           robotName={robotName}
           runId={runId}
         />
@@ -134,6 +141,10 @@ export function ProtocolRunSetup({
       description: t(`${MODULE_SETUP_KEY}_description`, {
         count: modules.length,
       }),
+    },
+    [LPC_KEY]: {
+      stepInternals: (<SetupLabwarePositionCheck {...{runId, robotName}} />),
+      description: t('labware_position_check_step_description'),
     },
     [LABWARE_SETUP_KEY]: {
       stepInternals: (
