+++ app/src/pages/OnDeviceDisplay/ProtocolSetup.tsx
@@ -352,16 +352,16 @@ function PrepareToRun({
           status={modulesStatus}
         />
         <ProtocolSetupStep
+          onClickSetupStep={launchLPC}
+          title={t('labware_position_check')}
+          detail={t('optional')}
+          status="general"
+        />
+        <ProtocolSetupStep
           onClickSetupStep={() => setSetupScreen('labware')}
           title={t('labware')}
           detail={labwareDetail}
           subDetail={labwareSubDetail}
-          status="general"
-        />
-        <ProtocolSetupStep
-          onClickSetupStep={launchLPC}
-          title={t('labware_position_check')}
-          detail={t('optional')}
           status="general"
         />
         <ProtocolSetupStep
@@ -371,8 +371,8 @@ function PrepareToRun({
           detail={
             liquidsInProtocol.length < 0
               ? t('initial_liquids_num', {
-                  num: liquidsInProtocol.length,
-                })
+                num: liquidsInProtocol.length,
+              })
               : t('liquids_not_in_setup')
           }
         />
@@ -443,7 +443,7 @@ function ProtocolSetupSkeleton(props: ProtocolSetupSkeletonProps): JSX.Element {
         </Flex>
         <Flex gridGap={SPACING.spacing5}>
           <CloseButton onClose={() => props.cancelAndClose()} />
-          <PlayButton disabled onPlay={() => {}} />
+          <PlayButton disabled onPlay={() => { }} />
         </Flex>
       </Flex>
       <Flex flexDirection={DIRECTION_COLUMN} gridGap={SPACING.spacing3}>
